requirements.txt
flask==3.0.3
line-bot-sdk==3.12.0
requests==2.32.3
gunicorn==22.0.0
